Extract SDK executor connections facade
